Mandelbrot.jl

Package Features

Calculate things about the Mandelbrot set

Function Documentation

Mandelbrot.KneadingSequence — Type

A kneading sequence is a combinatorial description of the orbit of an angle under doubling. For an angle theta, the kneading sequence of theta is defined as the theta itinerary of theta.

Mandelbrot.BinaryExpansion — Type

The binary expansion of an angle is a sequence of 1s and 0s which can be interpreted either as place-value representation of the angle, or as an itinerary of the angle under angle doubling.

Mandelbrot.OrientedHubbardTree — Type

A combinatorial Hubbard tree with cyclic order at each vertex. It can be constructed from an AngledInternalAddress or a RationalAngle, and it can be used to construct a HyperbolicComponent

Mandelbrot.InternalAddress — Type

An internal address is an increasing sequence of integers which describes a kneading sequence

Mandelbrot.HubbardTree — Type

HubbardTree(K::KneadingSequence)

Creates a topological Hubbard tree from the given kneading sequence using the triod algorithm described in [1].

Mandelbrot.AngledInternalAddress — Type

An internal address with aditional information in the form of angles which specify a cyclic ordering of kneading sequences